refactor: command 패키지 경로 변경
