+++ src/main/java/com/programmers/application/controller/command/execution/VoucherExecution.java
@@ -1,4 +1,4 @@
-package com.programmers.application.domain.command.execution;
+package com.programmers.application.controller.command.execution;
 
 import java.io.IOException;
